backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
python-multipart==0.0.6
aiosqlite==0.19.0

# Database drivers
asyncpg==0.29.0  # PostgreSQL async driver

# AI Integration
openai==1.3.0  # Optional: For AI-powered definitions and examples

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
